Moving CustomDatePicker into it's own file.
