+++ frontend/src/pages/Today.jsx
@@ -1,28 +1,5 @@
 import React, { useState } from "react";
-import DatePicker from "react-datepicker";
-import "react-datepicker/dist/react-datepicker.css";
-import "./Today.css"; // Import your custom CSS file
-
-// Custom Header Component
-const CustomHeader = ({
-  date,
-  decreaseMonth,
-  increaseMonth,
-  prevMonthButtonDisabled,
-  nextMonthButtonDisabled,
-}) => (
-  <div className="custom-header">
-    <button onClick={decreaseMonth} disabled={prevMonthButtonDisabled}>
-      {"<"}
-    </button>
-    <div className="month-year-label">
-      {date.toLocaleString("default", { month: "long" })} {date.getFullYear()}
-    </div>
-    <button onClick={increaseMonth} disabled={nextMonthButtonDisabled}>
-      {">"}
-    </button>
-  </div>
-);
+import CustomDatePicker from "./CustomDatePicker";
 
 function Today() {
   const [startDate, setStartDate] = useState(new Date());
@@ -30,16 +7,13 @@ function Today() {
   return (
     <div>
       <div>Today</div>
+
       <div className="container">
         <div></div>
-        <DatePicker
-          selected={startDate}
+
+        <CustomDatePicker
+          selectedDate={startDate}
           onChange={(date) => setStartDate(date)}
-          inline
-          minDate={new Date()}
-          filterDate={(date) => date.getDay() !== 6 && date.getDay() !== 0}
-          calendarClassName="custom-calendar"
-          renderCustomHeader={CustomHeader}
         />
       </div>
     </div>
